+++ tests/conftest.py
@@ -8,7 +8,6 @@ from sqlalchemy.orm import scoped_session, sessionmaker
 from sqlalchemy.pool import NullPool
 from zope.sqlalchemy import register
 
-from pyramid.compat import text_type
 from pytest_pyramid import factories
 import pyramid_basemodel
 
@@ -83,22 +82,22 @@ def active_user(user, db_session):  # pylint:disable=redefined-outer-name
 @pytest.fixture(
     params=[
         # (an @ character must separate the local and domain parts)
-        text_type("Abc.example.com"),
+        "Abc.example.com",
         # (character dot(.) is last in local part)
-        text_type("Abc.@example.com"),
+        "Abc.@example.com",
         # (character dot(.) is double)
-        text_type("Abc..123@example.com"),
+        "Abc..123@example.com",
         # (only one @ is allowed outside quotation marks)
-        text_type("A@b@c@example.com"),
+        "A@b@c@example.com",
         # (none of the special characters in this local part is allowed outside quotation marks)
-        text_type('a"b(c)d,e:f;g<h>i[j\\k]l@example.com'),
+        'a"b(c)d,e:f;g<h>i[j\\k]l@example.com',
         # (quoted strings must be dot separated, or the only element making up the local-part)
-        text_type('just"not"right@example.com'),
+        'just"not"right@example.com',
         # (spaces, quotes, and backslashes may only exist when within quoted strings and preceded by a backslash)
-        text_type('this is"not\\allowed@example.com'),
+        'this is"not\\allowed@example.com',
         # (even if escaped (preceded by a backslash), spaces, quotes, and backslashes must still be contained by quotes)
-        text_type('this\\ still"not\\allowed@example.com'),
-        text_type("bad-mail"),
+        'this\\ still"not\\allowed@example.com',
+        "bad-mail",
     ]
 )
 def invalid_email(request):