+++ tests/unit/test_user_repr.py
@@ -1,5 +1,4 @@
 """Test user representation."""
-from pyramid.compat import text_type
 
 from pyramid_fullauth.models import User
 
@@ -7,7 +6,7 @@ from pyramid_fullauth.models import User
 def test_introduce_email():
     """User gets introduced by e-mail only."""
     user = User()
-    user.email = text_type("test@test.pl")
+    user.email = "test@test.pl"
     # To string should return concatenated email
     assert str(user) == "test@..."
 
@@ -21,8 +20,8 @@ def test_introduce_username():
     # if id is set, user should be presented with it
     assert str(user) == "1"
 
-    user.email = text_type("test@test.pl")
-    user.username = text_type("testowy")
+    user.email = "test@test.pl"
+    user.username = "testowy"
     # To string should return username
     assert str(user) == "testowy"
     # User should be represented by "<User (\'1\')>"